MrBiens edit code logic PurchaseInvoice

+++ src/main/java/com/vn/sbit/SpringMVC/dto/request/ProductSupplierRequest.java
@@ -14,6 +14,6 @@ public class ProductSupplierRequest {
     private Long id;
     private Long productId;
     private Long supplierId;
-    private Double purchasePrice;
+    private int purchasePrice;
 
 }
