+++ src/main/java/com/vn/sbit/SpringMVC/service/PurchaseDetailService.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 public interface PurchaseDetailService extends CRUDService<PurchaseDetailRequest, PurchaseDetailUpdateRequest, PurchaseDetailResponse> {
     public List<PurchaseDetailResponse> getPurchaseInvoiceDetailsByPurchaseInvoiceId(Long purchaseInvoiceId) ;
-    public Double calculateTotalAmountByInvoiceId(Long purchaseInvoiceId);
+    public Integer calculateTotalAmountByInvoiceId(Long purchaseInvoiceId);
     public PurchaseInvoiceDetail getById(Long purchaseDetailId);
 
     public Integer calculateTotalQuantityByInvoiceId(Long purchaseInvoiceId);
